增加 root，大大提升项目可玩性。

+++ nezha.sh
@@ -5,6 +5,11 @@ NEZHA_SERVER=$1
 NEZHA_PORT=$2
 NEZHA_KEY=$3
 
+# 检测是否已运行
+check_run() {
+  [[ $(ps aux) =~ nezha-agent ]] && echo "哪吒客户端正在运行中" && exit 0
+}
+
 # 三个变量不全则不安装哪吒客户端
 check_variable() {
   [[ -z "${NEZHA_SERVER}" || -z "${NEZHA_PORT}" || -z "${NEZHA_KEY}" ]] && exit 0
@@ -20,9 +25,11 @@ check_dependencies() {
 
 # 下载最新版本 Nezha Agent
 download_agent() {
-  URL=$(wget -qO- -4 "https://api.github.com/repos/naiba/nezha/releases/latest" | grep -o "https.*linux_amd64.zip")
-  wget -t 2 -T 10 -N ${URL}
-  unzip -qod ./ nezha-agent_linux_amd64.zip && rm -f nezha-agent_linux_amd64.zip
+  if [ ! -e nezha-agent ]; then
+    URL=$(wget -qO- -4 "https://api.github.com/repos/naiba/nezha/releases/latest" | grep -o "https.*linux_amd64.zip")
+    wget -t 2 -T 10 -N ${URL}
+    unzip -qod ./ nezha-agent_linux_amd64.zip && rm -f nezha-agent_linux_amd64.zip
+  fi
 }
 
 # 运行客户端
@@ -30,6 +37,7 @@ run() {
   ./nezha-agent -s ${NEZHA_SERVER}:${NEZHA_PORT} -p ${NEZHA_KEY}
 }
 
+check_run
 check_variable
 check_dependencies
 download_agent
